feat: bin/update script
feat: check for uncommited changes

feat: bin/update and fixes in bin/configure
